// src/app/Models/BaseSpecialFlowDataClass.ts
// Generated by https://quicktype.io

export class BaseSpecialFlowDataClass {
    SpecialFlowID: number;
    GroupName:     string;
    SourceEmpID:   string;
    SourceName:    string;
    TargetEmpID:   string;
    TargetName:    string;
    Sort:          number;
    IsValid:       boolean;
    Note:          string;
    UpdateMan:     string;
    UpdateDate:    string;
}