// src/app/Models/PostData_API_Class/UpdateBaseSpecialFlowGetApiClass.ts
// Generated by https://quicktype.io

export class UpdateBaseSpecialFlowGetApiClass {
    SpecialFlowID: number;
    SourceEmpID:   string;
    TargetEmpID:   string;
    Sort:          number;
    IsValid:       boolean;
    Note:          string;
    GroupName:     string;
    UpdateMan:     string;
}